feat(coll): 댓글 삭제 로직 변경
대댓글이 없을 경우(즉, 원댓글인데 대댓글이 없거나 대댓글일 경우 - depth 1 제한이라 대댓글엔 대댓글이 없음)에는 hard delete, 그 외에는 soft delete

+++ src/main/java/org/devkor/apu/saerok_server/domain/collection/application/CollectionCommentCommandService.java
@@ -147,6 +147,11 @@ public class CollectionCommentCommandService {
             throw new ForbiddenException("해당 댓글에 대한 삭제 권한이 없어요");
         }
 
-        commentRepository.remove(comment);
+        // 대댓글이 있는 경우 soft delete, 없으면 hard delete
+        if (commentRepository.hasReplies(commentId)) {
+            comment.softDelete();
+        } else {
+            commentRepository.remove(comment);
+        }
     }
 }
